requirements.txt
langchain
langchain-google-genai
langchain-groq
langchain-community
PyJWT
deepagents
rich
pillow
duckduckgo-search
pyautogui
mss
openai
pywinauto
python-dotenv
groq
# MCP (Model Context Protocol) support
mcp[cli]
langchain-mcp-adapters
# Backend
flask
flask-sqlalchemy
requests